Collector/requirements.txt
bs4==0.0.2
mariadb==1.0.11
requests==2.20.0
PyYAML==3.12
lxml